java tests corrected
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@1090 7451a71c-b6c8-49cc-8254-141fc5361860

+++ libjzmq/Jzmq.java
@@ -42,8 +42,8 @@ class Jzmq
      //  or IP address of the network interface to be used by the exchange.
      //  Optionally you can specify the port to use this way "eth0:5555".
      //  Queue ID is returned.
-     public native int createQueue (String queue, int scope, String nic, int hwm, 
-		int lwm, int swap_size);
+     public native int createQueue (String queue, int scope, String nic, long hwm, 
+		long lwm, long swap_size);
 
      //  Bind the exchange to the queue.
      public native void bind (String exchange, String name,
